Multiple grain sizes now make up totalHeight, no alteration is done in setNodes.

+++ ISMoLD.py
@@ -4,64 +4,68 @@
 import math
 import os
 import matplotlib.pyplot as plt
+import numpy as np
+from os import listdir
 
 from functions_for_ISMoLD import *
 
-try:
+outputfilesNr = len(listdir("ISMolD_outputdata/"))-1 #-1 since file starts at 'output0'
+if (outputfilesNr >= 0):
     os.system('rm ISMolD_outputdata/topography*.txt') ##remove old data files
-except:
-    pass
 
 yr2sec = 60*60*24*365.25
 dx= 1.e3       # lateral grid spacing (m)
-dy = 1         # vertical grid spacing (m)
+dy = 1         # vertical grid spacing (m)  Important: MUST BE 1(as integer)!
 imax= 101      # number of nodes
 totalHeight= list(range(imax+1))
 newHeight= list(range(imax+1))
 x=list(range(imax+1))
+sedIncrease = np.zeros(shape=(2,imax+1))
+newSedContent = np.zeros(shape=(2,imax+1))
 t= 0
 dt= 1          # time step in (yr)
-tmax= 100000*yr2sec     # max number of years
+tmax= 10000*yr2sec     # max number of years
 dtout = tmax/100     # nr of years between write output
 dtout_progress = 10*yr2sec     # nr of years between progress bar update
 tout = 0.          # threshold to write next output in years (increased every write action by dtout)
 tout_progress = 0.          # threshold to update progress bar
 tprogress = 0.
 dtprogress = tmax/1000. 
-k= list(range(3))
-k[0]= 3.2e-4       # Gravel diffusivity (m2/s)
-k[1]= 3.2e-3       # Sand diffusivity (m2/s)
-k[2]= 3.2e-3       # Silt diffusivity (m2/s)
-q0= list(range(3))
+k= list(range(2))
+k[0]= 3.2e-5       # Gravel diffusivity (m2/s)
+k[1]= 3.2e-5       # Sand diffusivity (m2/s)
+q0= list(range(2))
 q0[0]= 2.e-6      # Gravel input (m2/s)
-q0[1]= 1.e-10      # Sand input (m2/s)
-q0[2]= 0           # silt input (m2/s)
+q0[1]= 2.e-6      # Sand input (m2/s)
 
-rho0 = 2700
+rho0 = dy * 2700
 
 totalInput= 0
 totalOutput= 0
 subsidence_rate = 0.#4e-4*dt/3 #m/yr
 maxVolumeLoss = 0
 
-
-
 ## Initialize:
 columns = {}
 for i in range(imax+1):
     x[i]=i
     totalHeight[i] = 0
-    columns[i]= {"height":0,
-                 "nodes":{}
+    columns[i]= {"totalHeight":0,
+                 "TotalSedContent":list(range(2)),
+                 "nodes":{},
                  }
+    for p in range(2):
+        columns[i]["TotalSedContent"][p] = 0
+        newSedContent[p,i]= 0
+
 
 ## ## ## ## ## ## ## ##
 ##  Main time loop:  ##
 ## ## ## ## ## ## ## ##
 while (t < tmax):
-    
-    dt = 0.9*(dx*dx)/(2.e0*(max(k)))
-    q0[0] = max(0, 2.e-6*math.sin(3.1415 * 6* t/tmax))
+    #dt = 0.9*(dx*dx)/(2.e0*(max(k)))
+    dt = 0.9*(dx*dx)/(2.e0*(sum(k)))
+    #q0[0] = max(0, 2.e-6*math.sin(3.1415 * 6* t/tmax))
     
     #dt = min(tmax/1000, 0.9*(dx*dx)/(2.e0*(k[0])) )
     if t > tout_progress:
@@ -70,52 +74,75 @@ while (t < tmax):
     
     ## Set boundary condition at proximal end of the basin:
     
-    ## Important note! If columns[0]["height"] is changed here to accommodate the boundary condition, no exception has to be made in ...
-    ## ... the FTCS method for i==1 and i==0. BUT the old columns[0]["height"] is required in setNodes. So it would need some sort ...
+    ## Important note! If columns[0]["totalHeight"] is changed here to accommodate the boundary condition, no exception has to be made in ...
+    ## ... the FTCS method for i==1 and i==0. BUT the old columns[0]["totalHeight"] is required in setNodes. So it would need some sort ...
     ## ... of columns[0]["OLD_height"] attribute.
-    newHeight[0] = columns[0]["height"]+q0[0]*dt/dx
+    #newHeight[0] = columns[0]["totalHeight"]
+    for p in range(2):
+        #newHeight[0] = newHeight[0]+q0[p]*dt/dx
+        sedIncrease[p,0] = q0[p]*dt/dx
+        columns[0]["TotalSedContent"][p] += sedIncrease[p,0]
     
     ## The right end of basin always remains 0 for it is not altered by the FTCS scheme. Therefore there is no need to specify that boundary condition here.
     
     ## Keep track of volume balance:
-    totalInput+= q0[0]*dt
-    totalOutput+= ( (k[0]*dt)/(dx*dx) )*( columns[imax-1]["height"] - columns[imax]["height"] )*dx ## Volume leaving = height*dx at i=imax
-    
+    for p in range(2):
+        totalInput+= q0[p]*dt
+        totalOutput+= ( (k[p]*dt)/(dx*dx) )*( columns[imax-1]["totalHeight"] - columns[imax]["totalHeight"] )*dx ## Volume leaving = height*dx at i=imax, which is set to 0
+        print(0, columns[0]["totalHeight"], sedIncrease[p,0], columns[0]["TotalSedContent"][p])
     ## Loop through columns (for FTCS, density calculations, etc):
     for i in range(1,imax):
-        #Dph= ( D(i+1)+D(i) )/2.e0
-        #Dmh= ( D(i)+D(i-1) )/2.e0
-        Dph = k[0]
-        Dmh = k[0]
-        newHeight[i]= columns[i]["height"] + ( (Dph*dt)/(dx*dx) )*( columns[i+1]["height"] - columns[i]["height"] )
-        if (i==1):
-            ##Consider effect of proximal boundary condition (i.e. i==0):
-            newHeight[1]= newHeight[1] - ( (Dmh*dt)/(dx*dx) )*( columns[1]["height"] - columns[0]["height"]-q0[0]*dt/dx ) 
-        else:
-            newHeight[i]= newHeight[i] - ( (Dmh*dt)/(dx*dx) )*( columns[i]["height"] - columns[i-1]["height"] )
+        newHeight[i]= columns[i]["totalHeight"] ##Start from current height
+        newSedContent[p,i] = columns[i]["TotalSedContent"][p] ##Start from current sediment content
+        for p in range(2):
+            ## FTCS:
+            #Dph= ( D(i+1)+D(i) )/2.e0
+            #Dmh= ( D(i)+D(i-1) )/2.e0
+            Dph = k[p]
+            Dmh = k[p]
+            sedIncrease[p,i] = ( (Dph*dt)/(dx*dx) )*( columns[i+1]["totalHeight"] - columns[i]["totalHeight"] )
+            #newHeight[i]= newHeight[i] + ( (Dph*dt)/(dx*dx) )*( columns[i+1]["totalHeight"] - columns[i]["totalHeight"] )
+            if (i==1):
+                ##Consider effect of proximal boundary condition (i.e. i==0):
+                sedIncrease[p,1] = sedIncrease[p,1] - ( (Dmh*dt)/(dx*dx) )*( columns[1]["totalHeight"] - columns[0]["totalHeight"]-q0[p]*dt/dx ) 
+                #newHeight[1]= newHeight[1] - ( (Dmh*dt)/(dx*dx) )*( columns[1]["totalHeight"] - columns[0]["totalHeight"]-q0[p]*dt/dx ) 
+            else:
+                sedIncrease[p,i] = sedIncrease[p,i] - ( (Dmh*dt)/(dx*dx) )*( columns[i]["totalHeight"] - columns[i-1]["totalHeight"] )
+            print(i, columns[i]["totalHeight"], sedIncrease[p,i], columns[i]["TotalSedContent"][p])
+            sedIncrease[p,i] = min(sedIncrease[p,i], columns[i-1]["TotalSedContent"][p])
+            newSedContent[p,i] += sedIncrease[p,i]
+            newHeight[i] += sedIncrease[p,i]
         
         ##Initial spike test:
         #if (i==70 and t==0): 
             #newHeight[70] = 100
-            #for p in range(100):
-                #columns[70]["nodes"].update({p:{"density":rho0}})
+            #for q in range(100):
+                #columns[70]["nodes"].update({q:{"density":rho0}})
             #print(columns[70])
-        columns[i] = setNodes(i, k, newHeight[i], columns[i], dt, dx, dy, rho0)
+        columns[i] = setNodes(i, k, newHeight[i], columns[i], dt, dx, dy, rho0)  ## Update density, (todo grain size fractions and porosity)
         
+    ## End column loop (i)
+    
     ## Set nodes in proximal boundary column:
-    newHeight[0]= newHeight[0] - ( (Dph*dt)/(dx*dx) )*( columns[0]["height"]+q0[0]*dt/dx - columns[1]["height"] )
+    for p in range(2):
+        sedIncrease[p,0] = sedIncrease[p,0] - ( (k[p]*dt)/(dx*dx) )*( columns[0]["totalHeight"]+q0[p]*dt/dx - columns[1]["totalHeight"] )
+        sedIncrease[p,0] = max(sedIncrease[p,0], -columns[0]["TotalSedContent"][p])
+        columns[0]["TotalSedContent"][p] += sedIncrease[p,0]- q0[p]*dt/dx ## q0 is added before FTCS to both sedIncrease and "TotalSedContent". q0 should not be added twice so it is subtracted once here.
+        newHeight[0] += sedIncrease[p,0]
     columns[0] = setNodes(0, k, newHeight[0], columns[0], dt, dx, dy, rho0)
     
     ## Overwrite old the topography with the new profile:
     for i in range(0,imax):
-        columns[i]["height"] = newHeight[i]
+        columns[i]["totalHeight"] = newHeight[i]
+        for p in range(2):
+            columns[i]["TotalSedContent"][p] = newSedContent[p,i]
     
     
     if (t >= tout):
         n= int(tout/dtout)
         f = open("ISMolD_outputdata/topography"+str(n)+".txt", "w")
         for i in range(len(x)):
-            f.write(str(x[i])+" "+str(columns[i]["height"])+"\n")
+            f.write(str(x[i])+" "+str(columns[i]["totalHeight"])+"\n")
         f.close()
         #call writeOutput(n, imax, x, hn, bedrock, totalHeight)
         tout = tout + dtout
@@ -123,6 +150,7 @@ while (t < tmax):
     
     t += dt
     
+## End time loop
 
 ## ## ## ## ## ## ## ##
 ##    Wrapping up:   ##
@@ -131,8 +159,8 @@ while (t < tmax):
 totalDepositVolume= 0
 totalNodeVolume= 0
 for i in range(imax+1):
-    totalHeight[i]= columns[i]["height"]
-    totalDepositVolume+= columns[i]["height"]*dx
+    totalHeight[i]= columns[i]["totalHeight"]
+    totalDepositVolume+= columns[i]["totalHeight"]*dx
     for j in range(0,len(columns[i]["nodes"])):
         try:
             totalNodeVolume+= dx*dy * columns[i]["nodes"][j]["density"]/rho0
@@ -140,7 +168,7 @@ for i in range(imax+1):
             pass
     
     
-#print(columns[70])
+print(columns[1]["TotalSedContent"], sum(columns[1]["TotalSedContent"]), totalHeight[1])
 print("totalInput:", str(totalInput)+"m^3")
 print("totalOutput:", str(totalOutput)+"m^3")
 print("totalDepositVolume:", str(totalDepositVolume)+"m^3")
@@ -152,7 +180,7 @@ print("")
 print("Node volume error:", str(totalNodeVolume+totalOutput-totalInput)+"m^3")
 print("Node volume error in %:", str(100*(totalNodeVolume+totalOutput-totalInput)/totalInput)+"%")
 print("")
-print("Plotting...")
+
 #plt.plot(x, totalHeight)
 #plt.show()
